+++ eds-aic-20230930/core/Models/DescribeBucketsRequest.cs
@@ -1,0 +1,22 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Eds_aic20230930.Models
+{
+    public class DescribeBucketsRequest : TeaModel {
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>docx</para>
+        /// </summary>
+        [NameInMap("FileType")]
+        [Validation(Required=false)]
+        public string FileType { get; set; }
+
+    }
+
+}