+++ eds-aic-20230930/core/Models/InstanceHealerRequest.cs
@@ -1,0 +1,37 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Eds_aic20230930.Models
+{
+    public class InstanceHealerRequest : TeaModel {
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// </summary>
+        [NameInMap("InstanceIdList")]
+        [Validation(Required=false)]
+        public List<string> InstanceIdList { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>Clean</para>
+        /// </summary>
+        [NameInMap("Strategy")]
+        [Validation(Required=false)]
+        public string Strategy { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>30</para>
+        /// </summary>
+        [NameInMap("Timeout")]
+        [Validation(Required=false)]
+        public long? Timeout { get; set; }
+
+    }
+
+}